# Simplify GT Assembler recipes for Muffler Upgrades