~~Currently, Muffler Upgrades have 6 recipe variants which are a combination of one of 3 plates and one of 2 dusts, but they all give the same result despite some variants being more expensive. The table below shows the previous recipes and the changes made, the balancing has been done based on how hard to get each material is.~~

| Plate Material | Dust Material | Original Output | Original Duration    | New Output | New  Duration        |
| -------------- | ------------- | --------------- | -------------------- | ---------- | -------------------- |
| Aluminium      | Plastic       | 1               | 1 minute, 20 seconds | 4          | 1 minute, 20 seconds |
| Aluminium      | Wood          | 1               | 1 minute, 20 seconds | 2          | 1 minute             |
| Wrought Iron   | Plastic       | 1               | 1 minute, 20 seconds | 2          | 1 minute, 20 seconds |
| Wrought Iron   | Wood          | 1               | 1 minute, 20 seconds | 1          | 1 minute             |
| Iron           | Plastic       | 1               | 1 minute, 20 seconds | 1          | 1 minute             |
| Iron           | Wood          | 1               | 1 minute, 20 seconds | 1          | 1 minute, 20 seconds |

Edit:
Just remove all the recipes and leave the cheapest one with an updated recipe time to be faster. After all the muffler upgrade is a quality of life upgrade and the item isn't used in crafting anywhere else there is no point to have multiple recipe variants that give increasingly more.