@Alexdoru 2024-11-14
Pesronally I would of deleted all the recipes and kept the less expensive one only. This is not an item that you need in large quantities, there is no reason to have different recipes that increasingly become better.