@HFPTetraUro 2024-11-14
I agree with you that there are too many variants, but I do think the lowest cost recipe takes too long to make early game, and you would want dozens of it if you make setups with many machines, so I think there should be at least another variant which is faster/better yield. I just used the existing useless more expensive recipe variants for that.