@Alexdoru 2024-11-15
> I agree with you that there are too many variants, but I do think the lowest cost recipe takes too long to make early game, and you would want dozens of it if you make setups with many machines, so I think there should be at least another variant which is faster/better yield. I just used the existing useless more expensive recipe variants for that.

You're right the recipe takes too long, what about you keep only this one and make it 5 sec 30eu/t ?
![image](https://github.com/user-attachments/assets/3bd53e2d-2891-429a-b397-1c57f1e3a6cb)