@HFPTetraUro 2024-11-15
> > I agree with you that there are too many variants, but I do think the lowest cost recipe takes too long to make early game, and you would want dozens of it if you make setups with many machines, so I think there should be at least another variant which is faster/better yield. I just used the existing useless more expensive recipe variants for that.
> 
> You're right the recipe takes too long, what about you keep only this one and make it 5 sec 30eu/t ? ![image](https://private-user-images.githubusercontent.com/57050655/386411361-3bd53e2d-2891-429a-b397-1c57f1e3a6cb.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MzE2MzMwMzAsIm5iZiI6MTczMTYzMjczMCwicGF0aCI6Ii81NzA1MDY1NS8zODY0MTEzNjEtM2JkNTNlMmQtMjg5MS00MjlhLWIzOTctMWM1N2YxZTNhNmNiLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDExMTUlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQxMTE1VDAxMDUzMFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWM2M2M2MGIyMThhZjRmNGYzMDgwYWJhY2ZlZjE1MTUwYTE1NzI0OTllNGIyNzVmOWYxODk3YWM2YWFkN2Y5MDEmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.BJ4ij92C-4H8rl7i-SZ2_70tmvhOs4vSCBdw0wZYGSc)

That sounds fair. I'll commit tomorrow and run spotless too